Added back in reference to main.php on form
Main.php is where we want the data sent, but we do not want the browser
to actually display main.php, therefore we will have to write some form
of code that points the browser to the global.js file which can take the
parsed data from the form and put it in the proper syntax for
Highcharts.  Then we can display the generated chart directly there.